parse enumerateguilditem

+++ network/mhfpacket/msg_mhf_enumerate_guild_item.go
@@ -12,7 +12,7 @@ import (
 type MsgMhfEnumerateGuildItem struct {
 	AckHandle uint32
   GuildId uint32
-  Unk0 uint32
+  Unk0 uint16
 }
 
 // Opcode returns the ID associated with this packet type.
@@ -24,7 +24,7 @@ func (m *MsgMhfEnumerateGuildItem) Opcode() network.PacketID {
 func (m *MsgMhfEnumerateGuildItem) Parse(bf *byteframe.ByteFrame, ctx *clientctx.ClientContext) error {
   m.AckHandle = bf.ReadUint32()
   m.GuildId = bf.ReadUint32()
-  m.Unk0 = bf.ReadUint32()
+  m.Unk0 = bf.ReadUint16()
 	return nil
 }
